+++ mocap_popy/model_templates/rigidBodyTemplate.py
@@ -15,7 +15,7 @@ import mocap_popy.config.regex as regex
 from mocap_popy.model_templates.baseTemplate import BaseTemplate
 from mocap_popy.models.rigid_body import RigidBody
 
-LOGGER = logging.getLogger(__name__)
+LOGGER = logging.getLogger("RigidBodyTemplate")
 
 
 class RigidBodyTemplate(BaseTemplate):
@@ -63,8 +63,7 @@ class RigidBodyTemplate(BaseTemplate):
         return self
 
     def add_symmetry_prefix(self, side: str):
-        """!Add a symmetry prefix to the name, markers and segments in the template."""
-        self.name = f"{side}_{self.name}"
+        """!Add a symmetry prefix to the markers and segments in the template."""
         self.markers = [f"{side}_{marker}" for marker in self.markers]
 
         sided_segments = []